Function reference

Macro related

NiceNumbers.nice — Function

nice(x, mod = @__MODULE__)

If x is an expression it replaces all occuring numbers by NiceNumbers.

If x is a number it turns it into a NiceNumber.

Otherwise it does nothing.

NiceNumbers.@nice — Macro

@nice

Return equivalent expression with all numbers converted to NiceNumbers.

Miscellaneous Helper Functions

NiceNumbers.isrational — Function

isrational(n::NiceNumber)

Whether n is purely rational, i.e. has no square root portion.

NiceNumbers.reduce_root — Function

reduce_root(coeff, radicand)

Simplifies the expression $\textrm{coeff}\cdot\sqrt{\textrm{radicand}}$ by factoring all remaining squares in radicand into coeff.

NiceNumbers.nthroot — Function

nthroot(m::NiceNumber, n)

Returns the nth root of m. Works by repeatedly determining the square root and thus only for powers of two.

Index

NiceNumbers.isrational

NiceNumbers.nice

NiceNumbers.nthroot

NiceNumbers.reduce_root

NiceNumbers.@nice